# pulsebs_frontend/src/Assets/Animations/server/api/openapi.yaml
openapi: 3.0.1
info:
  title: Pandemic University Lecture Seat Booking System (PULSeBS)
  description: A system that allow students to book a seat for in presence lectures
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: /
...
paths:
...
  /api/lectures:
    get:
      tags:
      - lectures
      description: Get a filtered list of lectures
      operationId: apiLecturesGET
      parameters:
      - name: courseId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/lectures'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      x-swagger-router-controller: Lectures
...
components:
  schemas:
...
    lecture:
      required:
      - courseId
      - date
      - lectureId
      - mode
      - teacherId
      - time
      type: object
      properties:
        lectureId:
          type: integer
        courseId:
          type: string
        teacherId:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
          format: time
        mode:
          type: string
        room:
          type: string
        maxSeats:
          type: integer
      example:
        date: 2000-01-23
        mode: mode
        teacherId: 6
        maxSeats: 1
        time: time
        courseId: courseId
        room: room
        lectureId: 0
    lectures:
      type: array
      items:
        $ref: '#/components/schemas/lecture'
...